create a single and multi fallback font family file
